Move Valuetype wrapper class to language-server
