+++ libs/execution/src/lib/types/io-types/table.ts
@@ -2,11 +2,10 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-import { IOType } from '@jvalue/jayvee-language-server';
+import { IOType, Valuetype } from '@jvalue/jayvee-language-server';
 
 import { SQLColumnTypeVisitor } from '../../visitors/sql-column-type-visitor';
 import { SQLValueRepresentationVisitor } from '../../visitors/sql-value-representation-visitor';
-import { Valuetype } from '../valuetypes/valuetype';
 
 import { IOTypeImplementation } from './io-type-implementation';
