+++ libs/execution/src/lib/types/valuetypes/index.ts
@@ -1,10 +1,1 @@
-// SPDX-FileCopyrightText: 2023 Friedrich-Alexander-Universitat Erlangen-Nurnberg
-//
-// SPDX-License-Identifier: AGPL-3.0-only
-
-export * from './primitive';
-export * from './visitors';
-
-export * from './atomic-valuetype';
-export * from './valuetype';
-export * from './valuetype-util';
+export * from './standard-representation';
