+++ libs/execution/src/lib/visitors/sql-column-type-visitor.ts
@@ -2,9 +2,10 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-// eslint-disable-next-line import/no-cycle
-import { AtomicValuetype } from '../types';
-import { ValuetypeVisitor } from '../types/valuetypes/visitors/valuetype-visitor';
+import {
+  AtomicValuetype,
+  ValuetypeVisitor,
+} from '@jvalue/jayvee-language-server';
 
 export class SQLColumnTypeVisitor extends ValuetypeVisitor<string> {
   override visitBoolean(): string {
