+++ libs/language-server/src/lib/ast/wrappers/index.ts
@@ -6,3 +6,4 @@ export * from './util';
 
 export * from './ast-node-wrapper';
 export * from './cell-range-wrapper';
+export * from './value-type';
