+++ libs/language-server/src/lib/ast/wrappers/value-type/visitors/index.ts
@@ -2,5 +2,6 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
+// eslint-disable-next-line import/no-cycle
 export * from './valuetype-visitor';
 export * from './visitable-valuetype';
